+++ App.js
@@ -1,72 +1,6 @@
-import React, { useState } from "react";
-import {
-  StyleSheet,
-  Text,
-  View,
-  Button,
-  TextInput,
-  ScrollView,
-  FlatList,
-} from "react-native";
-import GoalItem from "./components/goals/GoalItem";
-import GoalInput from "./components/goals/GoalInput";
+import React from "react";
+import HomepageBasics from "./components/goals/HomepageBasics";
 
 export default function App() {
-  const [courseGoals, setCourseGoals] = useState([]);
-  const [isAddMode, setIsAddMode] = useState(false);
-
-  const addGoalHandler = (goalTitle) => {
-    // console.log(goalTitle);
-    setCourseGoals((courseGoals) => [
-      ...courseGoals,
-      { id: Math.random().toString(), value: goalTitle },
-    ]);
-    setIsAddMode(false);
-  };
-
-  const removeGoalHandler = (goalId) => {
-    setCourseGoals((courseGoals) => {
-      return courseGoals.filter((goal) => goal.id !== goalId);
-    });
-  };
-
-  const onCanelhandler = () => {
-    setIsAddMode(false);
-  };
-
-  return (
-    <View style={styles.screen}>
-      <Button
-        title="Add New Goals"
-        onPress={() => {
-          setIsAddMode(true);
-        }}
-      />
-      <GoalInput
-        goalHandler={addGoalHandler}
-        visible={isAddMode}
-        onCancel={onCanelhandler}
-      />
-      {
-        // this will give warning her as we dont have any key here and for that it expect certain format of data with key
-      }
-      <FlatList
-        keyExtractor={(item, index) => item.id}
-        data={courseGoals}
-        renderItem={(itemData) => (
-          <GoalItem
-            id={itemData.item.id}
-            onDelete={removeGoalHandler}
-            title={itemData.item.value}
-          />
-        )}
-      />
-    </View>
-  );
+  return <HomepageBasics />;
 }
-
-const styles = StyleSheet.create({
-  screen: { padding: 50 },
-});
-
-// add some styling to the view
