+++ components/goals/HomepageBasics.js
@@ -1,0 +1,70 @@
+import React, { useState } from "react";
+import {
+  StyleSheet,
+  Text,
+  View,
+  Button,
+  TextInput,
+  ScrollView,
+  FlatList,
+} from "react-native";
+import GoalInput from "./GoalInput";
+import GoalItem from "./GoalItem";
+
+export default function HomepageBasics() {
+  const [courseGoals, setCourseGoals] = useState([]);
+  const [isAddMode, setIsAddMode] = useState(false);
+
+  const addGoalHandler = (goalTitle) => {
+    // console.log(goalTitle);
+    setCourseGoals((courseGoals) => [
+      ...courseGoals,
+      { id: Math.random().toString(), value: goalTitle },
+    ]);
+    setIsAddMode(false);
+  };
+
+  const removeGoalHandler = (goalId) => {
+    setCourseGoals((courseGoals) => {
+      return courseGoals.filter((goal) => goal.id !== goalId);
+    });
+  };
+
+  const onCanelhandler = () => {
+    setIsAddMode(false);
+  };
+
+  return (
+    <View style={styles.screen}>
+      <Button
+        title="Add New Goals"
+        onPress={() => {
+          setIsAddMode(true);
+        }}
+      />
+      <GoalInput
+        goalHandler={addGoalHandler}
+        visible={isAddMode}
+        onCancel={onCanelhandler}
+      />
+      {
+        // this will give warning her as we dont have any key here and for that it expect certain format of data with key
+      }
+      <FlatList
+        keyExtractor={(item, index) => item.id}
+        data={courseGoals}
+        renderItem={(itemData) => (
+          <GoalItem
+            id={itemData.item.id}
+            onDelete={removeGoalHandler}
+            title={itemData.item.value}
+          />
+        )}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  screen: { padding: 50 },
+});
